framework/backend/python/packages/k1s0-consensus/pyproject.toml
[project]
name = "k1s0-consensus"
version = "0.1.0"
description = "Distributed consensus for k1s0 Python services: leader election, distributed locks, saga orchestration"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "prometheus-client>=0.20.0",
    "pyyaml>=6.0",
    "k1s0-error",
    "k1s0-config",
    "k1s0-db",
    "k1s0-domain-event",
    "k1s0-observability",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/k1s0_consensus"]
